+++ ChangeLog.md
@@ -2,4 +2,4 @@
 
 ## 1.0
 
-Initial version
+- NEW - Add "voyage" object + transport mode dictionary + picto - *16/08/2023* - 1.1